LayoutTests/http/tests/site-isolation/inspector/console/message-from-worker.html

<script src="/inspector/resources/inspector-test.js"></script>
<script>
    // Utilities for page code.
    function addIFrame(url, testId) {
        const iframe = document.createElement("iframe");
        iframe.src = url;
        document.body.appendChild(iframe);
    }

    function test() {
        // The purpose of these tests:
        // Web Inspector should receive console messages generated from workers correctly.

        const suite = InspectorTest.createAsyncSuite("MessageFromWorker");

        // Utilities for test code.
        let callback = () => {};
        WI.consoleManager.addEventListener(WI.ConsoleManager.Event.MessageAdded, (event) => {
            InspectorTest.log(`Got message: ${event.data.message._messageText}`);
            callback();
        });

        function runMessageTest(resolve, expected, func) {
            let seen = 0;

            callback = () => {
                if (++seen === expected)
                    resolve();
            };

            InspectorTest.evaluateInPage(func);
        }

        suite.addTestCase({
            name: "MessageFromWorker.WorkerInSamePage",
            description: "console message coming from worker in the page",
            test: (resolve) => runMessageTest(resolve, 2, `
                const worker1 = new Worker('resources/worker.js');
                worker1.postMessage('Hello');
            `),
        });

        suite.addTestCase({
            name: "MessageFromWorker.WorkerInSameOriginIFrame",
            description: "console message coming from worker in same origin iframe",
            test: (resolve) => runMessageTest(resolve, 3, `addIFrame('resources/worker-in-iframe.html', 'same-origin')`),
        });

        // FIXME <rdar://143596746>: Additional work needed around presenting cross-origin workers.
        /* suite.addTestCase({
            name: "MessageFromWorker.WorkerInCrossOriginIFrame",
            description: "console message coming from worker in cross origin iframe",
            test: (resolve) => runMessageTest(resolve, 3, `addIFrame('http://localhost:8000/inspector/console/resources/worker-in-iframe.html', 'cross-origin')`),
        }); */

        suite.runTestCasesAndFinish();
    }
</script>

<body onload="runTest()"></body>
